requirements.txt
streamlit
pypdf
sentence-transformers
faiss-cpu
langchain-community
langchain-text-splitters
numpy